Gradebooks

Documentation for Gradebooks.

Gradebooks.ATTENDANCE_LIMIT

Gradebooks.ATTENDANCE_PENALTY

Gradebooks.COURSE_CREDITS

Gradebooks.COURSE_POINT_DECIMALPLACES

Gradebooks.COURSE_POINT_SCALE

Gradebooks.COURSE_POINT_SCALE_PLUSMINUS

Gradebooks.COURSE_POINT_SYSTEM

Gradebooks.COURSE_QUALITYPOINTS_A

Gradebooks.COURSE_QUALITYPOINTS_APLUS

Gradebooks.COURSE_QUALITYPOINTS_PLUSMINUS

Gradebooks.DATE_FORMAT

Gradebooks.DAYSOFWEEKSYMBOLCODES

Gradebooks.DAYSYMBOLCODEMAP

Gradebooks.DISPLAY_CREDITS

Gradebooks.INSTITUTION_EMAILDOMAIN

Gradebooks.INSTITUTION_NAME

Gradebooks.MIDNIGHT

Gradebooks.MWF

Gradebooks.STRING_MATCH_THRESHOLD

Gradebooks.TR

DataFrames.DataFrame

Gradebooks.Assignment

Gradebooks.AttendanceRecord

Gradebooks.AttendanceStatus

Gradebooks.Class

Gradebooks.Course

Gradebooks.Credit

Gradebooks.Evaluation

Gradebooks.Grade

Gradebooks.GradeLevel

Gradebooks.Gradebook

Gradebooks.Instructor

Gradebooks.LetterGrade

Gradebooks.Mark

Gradebooks.Percent

Gradebooks.Person

Gradebooks.Point

Gradebooks.Question

Gradebooks.Roster

Gradebooks.Rubric

Gradebooks.Score

Gradebooks.Student

Gradebooks.StudentIndex

Gradebooks.Submission

Gradebooks.Team

Gradebooks.Term

Gradebooks.pct

Gradebooks.pt

Gradebooks.Attendance

Gradebooks.Exam

Gradebooks.Homework

Gradebooks.Other

Gradebooks.Paper

Gradebooks.Presentation

Gradebooks.Project

Gradebooks.Quiz

Gradebooks._update

Gradebooks.attendance_record!

Gradebooks.attendance_record!

Gradebooks.attendance_status_map_from_string

Gradebooks.attendance_sync!

Gradebooks.attendance_update!

Gradebooks.credit2lettergrade

Gradebooks.expand!

Gradebooks.export_gradebook

Gradebooks.frequency2codesymbols

Gradebooks.get_emails

Gradebooks.get_student

Gradebooks.gpa

Gradebooks.grade

Gradebooks.gradelevel2qualitypoints

Gradebooks.gradelevel2string

Gradebooks.grades_post!

Gradebooks.grades_sync!

Gradebooks.islate

Gradebooks.isleaf

Gradebooks.latepenalty

Gradebooks.load_gradebook

Gradebooks.make_person_codename

Gradebooks.make_person_name

Gradebooks.parse_date

Gradebooks.parse_datetime

Gradebooks.parse_time

Gradebooks.post_grade!

Gradebooks.safe_datetime_stamp

Gradebooks.safe_datetime_stamp

Gradebooks.save_gradebook

Gradebooks.score

Gradebooks.string2codename

Gradebooks.string2uppercase_symbol

Gradebooks.string_sanitize

Gradebooks.view_assignment

Gradebooks.view_attendance

Gradebooks.view_gradebook

Gradebooks.ATTENDANCE_LIMIT — Constant

How many absences are allowed before penalties are applied.

Gradebooks.ATTENDANCE_PENALTY — Constant

Point deduction for so many absences exceeding ATTENDANCE_LIMIT.

Gradebooks.COURSE_CREDITS — Constant

Credit hours of course.

Gradebooks.COURSE_POINT_DECIMALPLACES — Constant

Number of decimal places to display.

Gradebooks.COURSE_POINT_SCALE — Constant

Point division between letter grades.

Gradebooks.COURSE_POINT_SCALE_PLUSMINUS — Constant

Point +/- deviation between letter grades.

Gradebooks.COURSE_POINT_SYSTEM — Constant

Total number of points a student can earn.

Gradebooks.COURSE_QUALITYPOINTS_A — Constant

Quality points of an "A" for calculating weighted GPA.

Gradebooks.COURSE_QUALITYPOINTS_APLUS — Constant

Quality points of an "A+" for calculating weighted GPA.

Gradebooks.COURSE_QUALITYPOINTS_PLUSMINUS — Constant

Quality point +/- devation for calculating weighted GPA.

Gradebooks.Credit — Type

Type union of Point and Percent for cleaner function signatures.

Gradebooks.DATE_FORMAT — Constant

Regional preference for whether month or day is written first in a date string by: "MMDDYYYY" (default) or "DDMMYYYY", respectively.

Gradebooks.DAYSOFWEEKSYMBOLCODES — Constant

Codes for each day of week. (Helpful for internal sorting.)

Gradebooks.DAYSYMBOLCODEMAP — Constant

Maps input to code representing day of the week.

Gradebooks.DISPLAY_CREDITS — Constant

Whether to display credit earned as Point by :point, Percent by :percent, or let the system decide with :auto (default).

Gradebooks.INSTITUTION_EMAILDOMAIN — Constant

Email domain: e.g., `"liberty.edu" or "mit.edu". This helps if searching by username instead of full email addresses.

Gradebooks.INSTITUTION_NAME — Constant

Institution name: e.g., "Liberty University" or "Massachussettes Institute of Technology".

Gradebooks.MIDNIGHT — Constant

Shorthand for 23:59:59.999.

Gradebooks.MWF — Constant

Shorthand for class frequency that meets Monday, Wednesday, and Friday.

Gradebooks.STRING_MATCH_THRESHOLD — Constant

Threshold for Levenshtein string matching.

Gradebooks.TR — Constant

Shorthand for class frequency that meets Tuesday and Thursday.

DataFrames.DataFrame — Method

Convert a vector of Student structs into a DataFrame, with one column per field.

Gradebooks.Assignment — Type

Is branch in course-assignment tree model always composed of at least one leaf question, even if not given.

is_group switches whether assignment should be completed individually.

Gradebooks.AttendanceRecord — Type

Records datetime stamp of attendance status[ with comments].

Gradebooks.AttendanceStatus — Type

Semantic options for how much attended of a class lecture.

Gradebooks.Class — Type

Couples course with prescribing assignments in term offered to attending student roster.

Gradebooks.Course — Type

An academic quest prescribing assignments to students to grow and evaluate their understanding and ability to think critically unto completing a degree.

credits defaults to COURSE_CREDITS preference.

Gradebooks.Evaluation — Type

Couples target question with redline mark.

Gradebooks.Grade — Type

Couples student to submitted work for assignment.

Gradebooks.GradeLevel — Type

Grade levels for predicate filtering and sorting.

Gradebooks.Gradebook — Type

Single source of truth recording all academic activity for class.

Gradebooks.Instructor — Type

Wrapper of Person with an academic job title.

Gradebooks.LetterGrade — Type

Letter grade (e.g., A or B+) earned for course item.

Gradebooks.Mark — Type

Encode redline marks of submitted work.

Resolves delta to Point unless explicitly given as Percent.

Gradebooks.Percent — Type

Percent earned of evaluation item.

Automatically divides by 100 so that basis is [0, 1].

Examples

julia> Percent(83)
Percent(0.83)

julia> Percent(0.83)
Percent(0.0083)

Gradebooks.Person — Type

Base structure for personal, contact, organizational, and metadata information.

Gradebooks.Point — Type

Points earned in the course.

Gradebooks.Question — Type

As the leaf in course-assignment tree model, can be nested for questions with parts or rubric categories.

Gradebooks.Roster — Type

Special type of Vector{Student}.

Gradebooks.Rubric — Type

Weight distributions to calculate grade from evaluations.

Gradebooks.Score — Type

Compares earned points to value with percent and letter grade.

Gradebooks.Student — Type

Wrapper of Person with an academic status unto a degree including final course performance.

Gradebooks.StudentIndex — Method

One time construction of students in class roster for fast, dictionary lookups of vector positions from any field of Person.

Gradebooks.Submission — Type

Couples datetime stamp of submitted work to computed score from evaluations.

Gradebooks.Team — Type

A group of students working as a team toward a common goal.

Gradebooks.Term — Type

Calendrical discretization of academic year.

Gradebooks.pct — Type

Shorthand construction with unit-like syntax: e.g., Percent(5.3) ≡ 5.3pct.

Gradebooks.pt — Type

Shorthand construction with unit-like syntax: e.g., Point(5.3) ≡ 5.3pt.

Gradebooks.Attendance — Function

Convenience function constructing Assignment to track attendance.

Gradebooks.Exam — Function

Convenience function constructing Assignment for an exam.

Gradebooks.Homework — Function

Convenience function constructing Assignment for homework.

Gradebooks.Other — Function

Convenience function constructing Assignment for some other academic item: e.g., extra credit.

Gradebooks.Paper — Function

Convenience function constructing Assignment for a paper.

Gradebooks.Presentation — Function

Convenience function constructing Assignment for a presentation.

Gradebooks.Project — Function

Convenience function constructing Assignment for a project.

Gradebooks.Quiz — Function

Convenience function constructing Assignment for a quiz.

Gradebooks._update — Method

Modifies the field(s) of an immutable struct and returns a new one.

Gradebooks.attendance_record! — Method

Records on date of entry (date_stamp) quality of attendance from a file matching regex within dir.

Warning

If using this method, make sure to implement attendance_status_map_from_string(x::AbstractString).

Gradebooks.attendance_record! — Method

Records quality of attendance for a class lecture (date_lecture) entered on date_stamp.

marks is magically parsed for combinations of AttendanceStatus, students, and comments.

Example

attendance_record!(gb, "Jun. 19, 2026", "Jun. 20, 2026", [NoRecord])
attendance_record!(gb, "2026-06-22", Date(2026, 06, 15), [
    (Excused, "Canceled due to snow")])
attendance_record!(gb, Date("2026-06-24"), Date("2026-06-25"), [
    ("alice@example.edu", Late, "Blew a tire"),
    (Present, "bob@example.edu", "Forgot to entire attendance code"),
    (Absent, "cole@example.edu"),
])

Gradebooks.attendance_status_map_from_string — Method

Maps string input to AttendanceStatus value.

Example

function attendance_status_map_from_string(x::AbstractString)
    s = lowercase(strip(x))
    return if s == "present"
        Present
    elseif s == "absent"
        Absent
    elseif s == "excused"
        Excused
    elseif s == "late"
        Late
    elseif s == "tardy"
        Tardy
    else
        missing
    end
end

Gradebooks.attendance_sync! — Method

Convenience function to ensure attendance penalties are current.

Gradebooks.attendance_update! — Method

Calculates number of Absent for each student and records penalties.

Gradebooks.credit2lettergrade — Method

Converts Credit to LetterGrade according to preferences COURSE_POINT_SCALE and COURSE_POINT_SCALE_PLUSMINUS.

Gradebooks.expand! — Function

Recursively applies marks down to the leaves of the appropriate branch.

Gradebooks.export_gradebook — Method

Exports gradebook (gb) to path in format (:json or :csv) which can also be parsed from file extension.

Gradebooks.frequency2codesymbols — Method

Magically maps input to vector subset of DAYSOFWEEKSYMBOLCODES.

Gradebooks.get_emails — Method

Gets emails from all students in roster. If search is not nothing, then returns a search string for email clients.

Gradebooks.get_student — Method

Find needle (identifier), according to any field of Person, in haystack (roster).

threshold adjusts Levenshtein string matching criterion.

Gradebooks.gpa — Function

Calculates weighted GPA from quality_points of LetterGrade and course credits (defaults to preference COURSE_CREDITS).

Gradebooks.grade — Method

Magically applies evaluations at any level of the course-assignment tree.

Gradebooks.gradelevel2qualitypoints — Method

Calculates quality point weight of LetterGrade according to preferences COURSE_QUALITYPOINTS_A, COURSE_QUALITYPOINTS_PLUSMINUS, and COURSE_QUALITYPOINTS_APLUS.

Gradebooks.gradelevel2string — Method

Maps grade level to string.

Gradebooks.grades_post! — Method

Post grades to the gradebook.

Can post grades for assignments from src searching for first non-empty row by column header.

Warning

This applies a scalar grade onto the entire assignment and risks erasing evaluation history in gradebook!

Gradebooks.grades_sync! — Method

Syncs current field value of gb.grades to gradebook.

Gradebooks.islate — Method

Helper function to determine whether submission was before assignment due date.

Gradebooks.isleaf — Method

Determines if a node is a leaf (i.e., has no children).

Gradebooks.latepenalty — Method

latepenalty(x)::Bool

Convenience function calling is_late and returning appropriate deductions for if late.

Will throw error if called and left unimplemented.

Example

Say the institutional policy for deductions of student work submitted late is as follows:

10% deduction if submitted within one week of due date

20% deduction if submitted within two weeks of due date

100% deduction if submitted two weeks or later of due date

Then this could be implemented by overwriting the Gradebooks.late_penaly(x::Millisecond) method:

julia> using Gradebooks
julia> Gradebooks.latepenalty(x::Millisecond) = Percent(is_late(x) ? (x < Day(7) ? 0.10 : (x < Day(14) ? 0.20 : 1.0)) : 0.0)
julia> latepenalty(Dates.CompoundPeriod(Day(1), Hour(12), Minute(36), Second(25)))
10%
julia> latepenalty(Dates.CompoundPeriod(Week(1), Day(1), Hour(12), Minute(36), Second(25)))
20%
julia> latepenalty(Dates.CompoundPeriod(Week(2), Day(1), Hour(12), Minute(36), Second(25)))
100%
julia> latepenalty(-Dates.CompoundPeriod(Week(1), Day(1), Hour(12), Minute(36), Second(25)))
0%

Gradebooks.load_gradebook — Method

Loads JLD2-serialized archive of GradebookArchive and returns gradebook field.

Gradebooks.make_person_codename — Method

Helper function to construct codename: e.g., make_person_codename(Joby, Anthony) ⟶ ":JA"

Gradebooks.make_person_name — Method

Helper function to construct name: e.g., make_person_name(Joby, Anthony; title="Dr.", suffix="III") ⟶ "Dr. Joby Anthony, III".

Gradebooks.parse_date — Method

Magically converts d to Dates.Date. Defaults to ISO 8601. (Uses DATE_FORMAT preference.)

Gradebooks.parse_datetime — Method

Magically converts dt to Dates.DateTime. Defaults to ISO 8601. (Uses DATE_FORMAT preference.)

Gradebooks.parse_time — Method

Magically converts t to Dates.Time. Defaults to ISO 8601.

Gradebooks.post_grade! — Method

Posts grade to gradebook.

Gradebooks.safe_datetime_stamp — Method

safe_datetime_stamp(dt::DateTime) = replace(string(dt), "-"=>"", ":"=>"", "."=>"")

Converts Dates.DateTime to a string safe for file paths.

Gradebooks.safe_datetime_stamp — Method

Safely append datetime stamp to file name.

Gradebooks.save_gradebook — Method

Saves gradebook (gb) to path with JLD2-serialization of GradebookArchive.

Gradebooks.score — Method

Resolves evaluations to total points earned which may be calculated from percent earned against assignment part value.

Gradebooks.string2codename — Method

Codifies string like one reads the title of a book but includes LaTeX-style exceptions with { and } pairs.

Gradebooks.string2uppercase_symbol — Method

Converts string to an uppercase symbol.

Gradebooks.string_sanitize — Method

Sanitizes a string by stripping all non-alphanumeric characters.

Gradebooks.view_assignment — Method

View detailed breakdown of assignment performance across entire class or subset according to index filter of students.

Gradebooks.view_attendance — Method

Similar to view_gradebook but specialized for attendance.

Gradebooks.view_gradebook — Method

View gradebook for entire class or subset according to index filter of assignments or students.